fix re-render of manage styles modal (sets)
