+++ packages/tokens-studio-for-figma/src/i18n/lang/en/manageStylesAndVariables.json
@@ -1,7 +1,7 @@
 {
 "modalTitle": "Export to Figma",
 "buttonLabel": "Styles and Variables",
-"optionsModalTitle": "Export to Figma Options",
+"optionsModalTitle": "Export to Figma: Options",
 "actions": {
   "export": "Export to Figma",
   "import": "Import from Figma",
@@ -14,7 +14,7 @@
   "changeSets": "Change Sets"
 },
 "docs": {
-  "exportToFigmaOptions": "Export to Figma Options",
+  "exportToFigmaOptions": "Export to Figma: Options",
   "stylesAndVariables": "Styles and Variables",
   "referenceOnlyMode": "Enabled versus Reference Only Mode"
 },
@@ -30,8 +30,9 @@
 },
 "options": {
   "title": "Options",
-  "whichVariablesShouldBeCreatedAndUpdated": "Which variables should be created and updated?",
-  "whichStylesShouldBeCreatedAndUpdated": "Which styles should be created and updated?",
+  "intro": "Create styles or variables based on your themes or token sets. First, choose what types of tokens should be created as either styles or variables.",
+  "whichVariablesShouldBeCreatedAndUpdated": "Create or update variables",
+  "whichStylesShouldBeCreatedAndUpdated": "Create or update styles",
   "tokensExportedToFigmaShould": "Tokens exported to Figma should:",
   "ignorePrefix": "Ignore first part of token name for styles",
   "ignorePrefixExplanation": "Useful if you want to ignore 'colors' in a token called 'colors.blue.500' for your styles",
@@ -56,18 +57,18 @@
   "effects": "Effects"
 },
 "exportThemesTab": {
-  "headingPro": "Create Themes to get Started",
-  "headingBasic": "Creating Themes is a pro Feature",
-  "introPro": "Combinations of token sets create themes. Use Themes to create multiple collections of variables or styles for easy switching between design concepts like light or dark color modes, multiple brands, products or platforms. Or switch to Sets to export without using Themes",
-  "introBasic": "Combinations of token sets create themes. Use Themes to create multiple collections of variables or styles for easy switching between design concepts like light or dark color modes, multiple brands, products or platforms.",
-  "confirmThemes": "Confirm Themes",
-  "combinationsOfSetsMakeThemes": "Combinations of token sets create themes"
+  "headingPro": "Create themes to get started",
+  "headingBasic": "Creating themes is a pro Feature",
+  "introPro": "Combinations of token sets create themes. Use themes to create multiple collections of variables or styles for easy switching between design concepts like light or dark color modes, multiple brands, products or platforms. Or switch to Sets to export without using Themes",
+  "introBasic": "Combinations of token sets create themes. Use themes to create multiple collections of variables or styles for easy switching between design concepts like light or dark color modes, multiple brands, products or platforms.",
+  "confirmThemes": "Create styles or variables based on themes",
+  "combinationsOfSetsMakeThemes": "Choose which themes should be created or updated."
 
 },
 "exportSetsTab": {
   "changeSetsHeading": "Enabled sets will export to Figma",
-  "confirmSets": "Confirm Sets",
-  "intro": "You can use token sets to manage your styles, and variables. Use Themes (Pro) to create multiple collections of variables or styles for easy switching between design concepts.",
+  "confirmSets": "Create styles or variables based on token sets",
+  "intro": "You can use token sets to manage your styles, and variables. Use themes (Pro) to create multiple collections of variables or styles for easy switching between design concepts.",
   "setsSelectedForExport": "Sets selected for export"
 }
 }
